A shorter alternative would have been https://commons.apache.org/io/api-release/org/apache/commons/io/FileUtils.html#forceMkdir%28java.io.File%29, but this works too.

+++ src/org/molgenis/generators/sql/HSqlCreateSubclassPerTableGen.java
@@ -3,6 +3,7 @@ package org.molgenis.generators.sql;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
 import java.sql.Connection;
@@ -47,7 +48,12 @@ public class HSqlCreateSubclassPerTableGen extends Generator
 
 		// Output file for debug
 		File target = new File(this.getSqlPath(options) + "/create_tables.sql");
-		target.getParentFile().mkdirs();
+		boolean created = target.getParentFile().mkdirs();
+		if (!created && !target.getParentFile().exists())
+		{
+			throw new IOException("could not create " + target.getParentFile());
+		}
+
 		OutputStream targetOut = new FileOutputStream(target);
 		List<Entity> sortedlist = model.getEntities();
 		sortedlist = MolgenisModel.sortEntitiesByDependency(sortedlist, model);
